fix: move componets in login page
